+++ routes/api/components.js
@@ -4,9 +4,11 @@ const path = require('path');
 const config = require('../../config.json')
 const { authCheck } = require('../../middleware/authentication/authentication');
 
+const { apiAuthCheck } = require('../../middleware/authentication/apiAuthentication');
+const { checkOrigin } = require('../../middleware/authentication/checkOrigin');
 
 // ! Components
-router.get('/:filename', authCheck, (req, res) => {    
+router.get('/:filename', authCheck, checkOrigin, (req, res) => {    
     res.render(`components/${req.params.filename}`, {
         query: req.query,
         config: config,